requirements.txt
boto3==1.26.145
botocore==1.29.145
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.16
